Finished task[T390]
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@2296 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ linux-2.6.x/drivers/usb/host/isp1362.h
@@ -736,7 +736,7 @@ static void isp1362_read_fifo(struct isp1362_hcd *isp1362_hcd, void *buf, u16 le
 	if (len >= 2) {
 		RDBG("%s: Using readsw for %d words\n", __FUNCTION__, len >> 1);
 #if 1		
-		insw(isp1362_hcd->data_reg, dp, len >> 1);
+		insw((unsigned long)isp1362_hcd->data_reg, dp, len >> 1);
 		dp += len & ~1;
 		len &= 1;
 #else
@@ -778,7 +778,7 @@ static void isp1362_write_fifo(struct isp1362_hcd *isp1362_hcd, void *buf, u16 l
 	if (len >= 2) {
 		RDBG("%s: Using writesw for %d words\n", __FUNCTION__, len >> 1);
 #if 1
-		outsw(isp1362_hcd->data_reg, dp, len >> 1);
+		outsw((unsigned long)isp1362_hcd->data_reg, dp, len >> 1);
 		dp += len & ~1;
 		len &= 1;
 #else
